Keep the current stream if the new one is not playable
Instead of killing the old process, we stop it and try to play the new one.
A sentinel will keep an eye on the new process:
- If the new process exits, that means that the target URL is not
playable by the media player. Warn the user and let the old process
play back the old stream.
- Otherwise, the old process will remain stopped until being killed
whenever the new steam changes status.

At any moment, eradio keeps at most two processes (one stopped and the other
running).

Some refactorings have also been made concerning processes. A benefit is
the reduced latency of eradio-toggle: we stop the process instead of
killing it, and continue it instead of restarting a connection.

Signed-off-by: Firmin Martin <firminmartin24@gmail.com>