Check whether the target URL is available
Sometimes URL may be broken. In this case, eradio-play will stop the stream being played and will silently be unable to play the newly selected stream. 
Fix this issue by checking the availability of the URL before doing any action aforementioned. If the URL is not available, warn the user and do not interrupt the current stream.